requirements.txt
kafka-python==2.0.2
pandas==1.3.5
s3fs==2023.1.0